requirements.txt
requests
pyyaml
lxml.html
